add option to pass vault lincese as a string
Having a license file on a host is not a viable option for everyone.

By passing the license as a string, then it can be stored as a secret in
Ansible or retrieved from an external data source.

This adds a new option without removing the existing `vault_license_file`
functionality.

To simplify the code, this commit changes the existing tasks to now use
`vault.hclic` for the license filename, regardless of the method used to
supply the license.