add option to pass TLS certs as strings
Having a sensitive files on a host where you run Ansible from is not a viable option for everyone.

By passing the the TLS certificates as strings, then they can be stored as secrets in
Ansible or retrieved from an external data source.

This adds a new option without removing the existing `vault_tls_*_file`
functionality.

To simplify the code, this commit changes the existing tasks to now use use hardcoded names for the TLS files.